requirements.txt
pandas
openai
matplotlib
seaborn
scikit-learn
python-dotenv